fix bug 1725
configure has been taught to gracefully handle systems missing a C++ compiler. In those cases, udp-conduit will now be automatically disabled and configure will continue as if the user had passed --disable-udp (except if they explicitly forced udp conduit with --enable-udp, in which case it drops dead).

Because AC_PROG_CXX provides no mechanism for failure recovery, the trick currently requires running a portion of the C++ compiler detection step twice on systems that have C++ (once in a subshell to ensure it won't drop dead, and a second time to actually get the results). This adds a small amount of time to configure, but I don't see any other way to get the desired effect (and it seems to be worthwhile since UPC users appear to run into it rather frequently).

We now also accept --without-cxx as an alias that disables udp-conduit.

CVSMAIL: CC: