Fix bug 1857 - vapi crashes on sunc/Linux/ndebug
I had no luck using compiler flags to set the most conservative alias
assumptions, so this appears to be something different from the errors
that have cropped up on this code w/ other compilers.
By adding a compiler fence in the right spot, the problem goes away.
I hate fixes that I can't explain, but this seems to be good enough.
CVSMAIL: CC: