Fix an unreported vapi-conduit bug.
The bounds check and automatic adjustment of gasnetc_inline_limit against
the HCA capabilities was being done after its previous value had already been
used to precompute some other bounds.

This would only create problems if the GASNET_INLINESEND_LIMIT environment
variable was set larger than the HCA capability, or if the HCA capability
was lower than the compiled-in default of 72 bytes (as is the case w/ the
OpenIB/OFED stack where the limit is 42).