fix bug 1947 - crashes, hangs and memory-grabbing on testtoolscxx
The gasnet_threadkey code for TLS had some contingency code for handling compiler mismatches (which includes use of a C++ compiler on the GASNet headers) where we cannot rely upon TLS working, but that code assumed the mismatch compiler would only be used to access threadkeys, but not define them. testtoolscxx actually defines a threadkey, and the code as written caused an infinite stack recursion upon access to that key (leading to the observed memory gobbling and seg faults on systems with TLS).

Fixed by adding a case for mismatched compilers defining a threadkey in a TLS-enabled configure, in which case we fall back upon a pthread_getspecific-based implementation (which may be suboptimal but is guaranteed to be safe). Key accesses under mismatched compiler to keys defined by the supported compiler will continue to use an extern function call and TLS support.

CVSMAIL: CC: