Mods to gasnetc_MaxPinMem, which uses a bisection algorithm to detect the max amount of memory on the node that can be pinned (registered as a portals memory descriptor). Unfortunately, under CNL, an attempt to pin more memory than the OS will allow causes the job to be killed. Since we have no non-destructive way to detect the max pinnable memory, we are effectively setting MaxLocalSegmentSize and MaxGlobalSegmentSize to infinity: (uintptr_t)-1. If the user requests too large of a segment, the code will die at run-time with an error similar to the following:
[NID 18]Apid 49229: initiated application termination
Application 49229 exit signals: Killed

which is not very helpful.

NOTE: I have discovered, that on Franklin, if you run one process per dual-core
node, then the bisection pinning algorithm actually works, the attempt to pin
returns an error rather than killing the job.  However, if you run two processes
on the dual-core node, the Linux Out-Of-Memory daemon kills the job.
In either case, the bisection algoritm gets slow when attempting to pin large
memory segments and we would not want to run this on every job launch.

CVSMAIL: CC: