Fix bug 2072 (Implement gasnett_getPhysMemSz() for IRIX) Using getinvent() family of routines, as suggested by page linked from bug report.
Online manpages from IRIX 5.3 show the interface has been around at least that long, and contents of the headers suggest 5.2 or earlier (makes reference to "post 5.2" additions).
Since the "class" and "type" tokens are not documented other than in the header, I've programmed defensively to avoid doing anything if the ones we need are not defined.
Since the traversal of the inventory is potentially expensive, we traverse it only once and cache the results.  Care has been taken to use the _r-suffixed thread-safe getinvent_r() family of calls.  Memory barriers are not includes, since we cache only a single word-sized quantity.