Fix generic atomic code to avoid excessive warnings from gcc 4.2.x Previous code got *many* warnings of the form  warning: passing argument 1 of 'gasneti_pthread_atomic_hash_lookup' discards qualifiers from pointer target type or  warning: passing argument 1 of 'gasneti_hsl_atomic_hash_lookup' discards qualifiers from pointer target type
Since we only hash on the address, we couldn't care less if the pointer we
pass is to a volatile object or not, but gcc-4.2.1 certainly does care.
The cast we applied to (void *) was not sufficient to rid us of this warning,
so an argument type of "uintptr_t" is now used.  Use of (const pointer to volatile
void) was another possibility, but I wonder if all compilers are prepared to deal
with that one.  (What *is* a "volatile void"?)