fix bug 2083: gm-conduit: intermittent bad payload delivery for AM Longs
gm-conduit has 3 separate possible transfer protocols for AMLong(Async) operations, depending on the async status and registration of the payload src/dst. When both are pinned and the operation is async, it tries to use zero-copy RDMA. When the remote is pinned but the local cannot be or is synchronous, the payload is copied to local bounce buffers and RDMA'd from there. Finally when the remote side is not pinned, the payload transfer is achieved by a series of what amount to two-copy AMMedium's that transfer the payload, followed by an header message to run the Long handler. Due to bug 2101, all AMLong(Async) operations in FAST mode are always using that third case.

The bug arose for PAR mode in the third case (only) because there was no synchronization between the AMMediums and the header message, and gm-conduit allows concurrent handler execution. Consequently a thread could come along, pickup the header message and run the AMLong handler, while a thread that polled earlier was still busy copying the data into place from one of the AMMedium payload packets.

The fix is to disable concurrent handler execution (by keeping the GM lock held) while processing the AMMediums for this special case. This is safe because these particular handlers only perform memcpy and never reply. This fix is somewhat suboptimal because it reduces concurrency with respect to unrelated handlers (eg from other nodes) that could otherwise be running on separate local threads while this payload copying takes place (esp if it takes a page fault), but algorithms to allow that concurrency while still maintaining the required ordering are non-trivial and would require significant changes to gm-conduit and possibly the packet format. Even with such a fix, this transfer mechanism would remain suboptimal from a bandwidth perspective (because it's inescapably two-copy), so the better performance fix is simply to fix bug 2101 to reduce the situations in which we use this fallback.

Fix has been tested successfully on CITRIS using the gasnet tests for FAST/LARGE/EVERYTHING, debug and opt, SEQ and PAR.

Also add some stats to count FAST vs NORMAL recv events for gm_receive
CVSMAIL: CC: