fix bug 2102: testcore2-par crashes on portals/CNL
The evidence indicates  that the implementation of PtlEQAlloc/PtlEQFree is very likely not thread-safe, and the frequent creation and destruction of EQ's on concurrent handler threads within AMReplyLong eventually results in a collision and a crash. This explanation is consistent with all the evidence of this problem to date, and I've confirmed that adding a new mutex to enforce a critical section directly around the relevant calls to PtlEQAlloc and PtlEQFree is sufficient to make the crashes disappear. I'm committing that workaround to resolve this issue for the purposes of our release.

CVSMAIL: CC: