more aggressive tester for collectives
this tester beats on the collectives at larger data sizes and many iterations. It also tests all 9 variants of the synch flags and their associated performance. The test is currently designed to use at most half of the gasnet segments and adjusts the users arguments until this is the case
CVSMAIL: CC: