fixed bug 2130
we were incorrectly assuming an op was finished even though we had an
outstanding NBI handle for that op that had yet to be synched. The
collectives were by passing the synch to the nbi handle when there were no
comm ops inside that region. There were a couple of other use after free
issues that i caught when scanning through the code fixing this bug.
CVSMAIL: CC: