fix bug 2181
The fix is similar to the approach used by mpi-conduit, namely we repeatedly try_lock the AMLOCK in a yield loop to try and acquire it (and lock out other threads) for exit-time processing. I've also added a flag variable which causes other threads to yield when acquiring the lock if some thread is attempting to acquire for exit. The udp-conduit case is somewhat complicated for GASNET_TRACE mode by the additional stats we dump to the trace log, so I've also modified the locking in that code to prevent other threads from sneaking in during that phase of exit.

With the fixes, udp-conduit now reliably passes all cases of testexit in seq and par mode, opt and debug with tracing off. With tracing on, a few of the tests (eg test 18) still occasionally cause an exit-time error message about reentrance in AMUDP_SPMDShutdown, but that doesn't seem likely to affect end users in practice.
CVSMAIL: CC: