Fix bug 2183 (11th hour commit agreed upon by Paul, Dan and Parry).
Testing on bdev shows no gasnet test failures when AM loopback is re-enabled.
My testing includes "make run-tests" for all eight combinations
{seq,par}X{opt,dbg}X{TEST_NODES=2,TEST_NODES=1}.
All tests are SEGMENT_FAST since there is no segment-dependent code to consider here.

Parry confirms that this change was not intentional.
Parry, Dan and I have spoken on the phone and agreed to include this fix in the release.