Fix bug 2257 (SEGVs at exit when firehose disabled) This commit replaces the broken freelist implementation described in bug 2257 with a semaphore.  The semaphore limits the number of in-flight PVOs as the freelist previously would have done, but there is no longer any need for the struct the freelist managed.
What I found is that the small struct that was being managed contained
just the user_pvo and the corresponding length, and existed to pass this
pair through to the LAPI send completion handler (via the single void *
provided by LAPI for that purpose).  However, the length was only used
for accounting the number of bytes pinned, and that accounting was:
1) incorrect, because pin and unpin both INCremented it (unpin should DECrement)
2) unused - the code never read the value kept.

So, only the 64-bit user_pvo is used and that can be passed in the sinfo (void
*) field.  Of course that is only true for a 64-bit build, but that is all we
support for LAPI-RDMA.  I've made comments about what needs to be done it we
ever want/need to support 32-bit builds.