Do not stringify __VERSION__ when adding misc: configuration info.
When upcr is configured to use GCC/UPC as the front-end,
the platform detection logic adds the version of the front-end
to the misc: attrbute in the configuration information string.
The __VERSION__ variable is already "string-ified", therefore
the explicit call to _STRINGIFY() is unneeded, and caused a
syntax error due improper quotation.  This fix removes the
call to _STRINGIFY().