fix bug 2312 -- cross-configure-help.c gives wrong STACK_GROWS_UP
Turns out the cross-configure-help check was correct, but the sense was reversed, and apparently all the cross-compile scripts are for systems with downward-growing stacks. The UPCR stack overflow warning check (the only current client of this test) was coded in such a way that an incorrect value would simply result in the check being silently disabled.

Fixed the cross-configure-help check and the scripts.
CVSMAIL: CC: