Fix bug 2325 (GASNET_SUBCONFIGURE_EXTRACT broken w/ new autoconf)
This commit applies the patch attached to the bug report.
I've tested w/ autoconf version 2.13 and 2.61 (our oldest version, and the
last version to use sed in config.status), and with 2.62 and 2.63 (the first
version to use awk in config.status, and the most recent version found
at ftp.gnu.org).