Partial fix for bug 2331
For both PA-RISC and SPARC < V8+, we use a clever check-out approach to
implementing our atomics.  In the atomic_set() they use a "magic number"
to distinguish the case of an uninitialized value from an initialized one.
This is needed because attempting a "checkout" of the uninitialized location
might hang indefinately (if the "present" bit is absent it will never appear).

In the case of bug 2331, it appears that the lockperf test has used a given
section of memory, has free()ed it, and has allocated it again later.  While
NOT allocated some of the memory for the atomic got "scribbled" on, but the
"magic number" survived.  Unfortunately, the presence of the "magic number"
makes us believe we need to follow the checkout protocol.  When we do the
assertion fails because the value read is non-zero but the present bit is
missing.

There is no total fix for the reuse of memory.  If the "magic" is there but
the atomic word is zero, then all the atomic operations will hang trying to
complete a checkout that will never happen.

However, I am going to *reduce* the impact of this problem as follows.
I have removed the assertion in gasneti_checkout_32() and replace it with
equivalent checks in its callers except atomic_set.  In atomic_set, if the
"magic" is correct, but the "present" bit is missing then things will still
just work because we'll blindly write in the present bit, never noticing that
is was gone.

The trade off here, however, is that we may miss "random writes" to the atomic
that currently cause the assertion to trigger.

Note this commit makes the change to both PARISC and SPARC (pre v8+).