fix bug 2475 - the last LANG-related perl bug
The culprit is Perl's implicit UTF support on input streams, based on the LANG setting (man perlrun for -C). The read of the binary string TARBALL was benig converted to UTF and corrupted in the process.

I had LC_ALL=C in my environment, which disables all of Perl's fragile implicit UTF support and forces it to act sanely. I've modified perlstart to force the same, which should hopefully be the end of all our LANG-related perl bugs.

Our scripts don't even support the full ASCII character set (ie bug 425), and UTF is not something I think we should tackle (certainly not given the flaky state of Perl's support for it).

CVSMAIL: CC: