Implement --disable-auto-conduit-detect option to configure. This option changes the default for conduits from enabled to disabled. The option name is intentionally long and hard to remember.
This was suggested by Brad at Cray as one way to resolve the problems they
have with portals-conduit when building with --enable-segment-everything.
The correct solution for that problem, however, was to make configure
disable any conduit that is incompatible with the selected segment config.
However, I have for a long time wanted this option for cases when I need
to configure and build GASNet and/or UPCR to reproduce a specific bug with
just one conduit.  So, I've taken Brad's suggestion as the push I needed
to finally implement this relatively simple feature.