Merge GASNETI_EXTERN_UNLESS_BUILDING_TOOLS into GASNETI_TENTATIVE_EXTERN.
The former had been introduced just recently to help deal with bug 2117,
while the later had been introduced a while back to deal with the fact
that C++ does not support tentative definitions.  I have come to realize,
however, that the AIX+gcc problem described in bug 2117 is realy just a
milder form of lacking support for tentative definitions.  Thus there is
no need for two macros when the one will suffice.