Pass srcnode in spare Portals header bits of Long Payload Put
Bug 2581 raises the issue of a linear-time id->srcnode lookup operation
on each arriving AM.  While there are tradeoffs between time and (wire) space
to be made for the headers, there is a no-contest win for packing the srcnode
in otherwise unused (and thus zero incremental cost on the wire) header bits
when sending the Payload Put of a non-packed Long.  So, this commit does so.