Fix an unreported bug in portals-conduit's gasnetc_AMReplyLongM()
While looking a the AM code, I noticed that the Long Reply code was explicitly
"returning a send ticket" though the same construct on the Long Request
was not.  I confirmed that this was duplicating a ticket return that is
performed in the Event handler for the SEND_END event.  The result was that
the sent ticket count would grow by one for every Long Reply, effectively
eliminating it's desirable throttling effect.