If gcc/g++ on ia64 support -mtls-size=64, then add to C and CXX flags.
This resolves a problem observed by Intrepid with respect to use of __thread
both for gasnet-internal handling of thread-specific data, and for that of
static per-thread UPC data w/ gccupc.  The problem is that the default width
of the thread local storage (TLS) offset is too small to address all of the
data in some cases.  Since it is required that both the library and application
be built with this option, it is being added to the flags at configure time.