Apply the bug 2646 fix unconditionally (with out regard to what compiler is used). This change is motivated by the following comment I just made in Bugzilla:
I have tried to think about why/how this apparent "compiler bug" is arising,
and in particular why the problem is not deterministic.  I am concerned that it
might actually be a fundamental problem with our code and ANY sufficiently
aggressive (or merely perverse) optimizer.

When we have something like
  if (i_am_owning_thread) var++;
we expect that only the owning thread will actually write tovar.

However, if the compiler chooses to, this might become
  var += (i_am_owning_thread ? 1 : 0);
or
  var = i_am_owning_thread ? (var + 1) : var;
among other possibilities.

With muliple threads running, either of the code variants above could result in
exactly the sort of "missing increment" that we are seeing in the duplicated
sequence number.  This would also explain why a volatile qualifier appears to
fix the problem (by prohibiting any "extra" writes).