Fix for differences between DCMF libs on Intrepid and Surveyor.
Somewhere between the versions of the BG/P systems software installed on
ANL's two machines there was a change in libdcmf and/or libdcmfcoll.
The result is that to link gasnet progs on one machine requires
  -ldcmf.cnk -ldcmfcoll.cnk
and the other requires
  -ldcmfcoll.cnk -ldcmf.cnk
With the wrong ordering one gets undefined symbols at link time.
This commit changes the "for working DCMF configuration" probe to try the
link test with two changes:
1) reference a symbol in libdcmfcoll to trigger the undefined symbol problem
2) if the first attempt fails, try again w/ the link order swapped
The logic to swap the link order finds the first two words beginning with
"-ldcmf" without making any additional assumptions about what is in $DCMF_LIBS.

Some additional changes in this commit:
+ The link test tries the order appropriate to the newer s/w first in the
  hope of running just a little faster.
+ If we discard dcmf-conduit due to missing/invalid settings we now issue a
  message saying why.  Example:
     checking for additional valid DCMF directories... no (DCMF_SYS_LIBDIR setting is empty)
+ Corrected "LIB_DIR" to "LIBDIR" in zepto support (aka --enable-dcmf-linux)