Quote the mkstemp() call in gasnet/configure.in.
Per, http://www.gnu.org/software/autoconf/manual/m4/Mkstemp.html
mkstemp() happens to be a builtin M4 macro.  Left unquated, the
reference is replaced with a temp filename that uses the argument
as a template.  This ends up placing the string filenamexxxxxx
(where xxxxxx are random characters) into the test program,
instead of the string mkstemp(filename), *and* it creates
a temporary file of the same name in the gasnet source
direectory when Bootstrap is run.  The other downside is
that this mmap() configuration test would always fail due
to compilation errors.

Adding '[...]' to quote this mkstemp() reference
fixes the problem.