Allow a GASNet client to directly control which GCC-style attributes are used. This is intended to allow GASNet client code to be built with compilers other than the ones known to GASNet at configure time (CC, CXX and MPI_CC).  The new GASNETT_USE_GCC_ATTRIBUTE_* macros control GASNet's use of __attribute__ in its public headers.  The expectation is that the client determines these either by configure probes or by inspection of version numbers of a well- known compiler family.
The anticipated "user" of this mechanism is GUPC+UPCR for the case when
CC is not the gcc built from the same sources as upc.  This is an
attempt (perhaps sub-optimal) to address Bugs 2721 and 2722 in which
the user's upc code (compiled by GUPC) does not contain any attributes
(except for always_inline which we previously special cased).  In
particular the may_alias attribute in gasnet_extended_help.h is ommitted
leading to incorrect code generation with recent GUPC versions.

This is NOT intended for building of the GASNet libraries with a compiler
other than the ones given as CC, CXX and MPI_CC at configure time.