Use recent additions to probe version and attribute info from $CXX and $MPI_CC This is part of on-going work on bugs 2721 and 2722, and will allow the current "compiler-mismatch" and related "attribute-squaching" logic to become far mor intelligent (thus less fragile).
This goal is that gasnet_basic.h will recognize when the compiler is one of
these probed at configuration time and apply the correct attributes.

The GASNETT_USE_GCC_ATTRIBUTE_* macros added yesterday will continue to
have greatest precedence.  Only when one of these is undefined will the
compiler-matching logic kick in.  Note that ONLY if we are run without a
definition for a given GASNETT_USE_GCC_ATTRIBUTE_* macro AND don't
recognize the compiler (as $CC, $CXX or $MPI_CC) will we resort to the
"squashing" of attributes (individually, not as a whole).