Fix handling of "restrict" in the same spirit as done for __attribute__.
As with __attribute__ we have been "squashing" the restrict (or equivalent)
keyword if the complier in use did not match $CC used at configure time.
For instance
+ Squashed for $MPI_CC if was not the same as $CC
+ Always squashed for $CXX unconditionally
+ Squashed for GCCUPC unless $CC was the matching version of gcc.

We now probe $CC, $CXX and $MPI_CC individually for their "restrict"
support and use the info to implement GASNETI_RESTRICT correctly at
compile time.

The availability of restrict to qualify typedefs (as opposed to pointers
to typedefs) is also handled in the same manner.

As with __attribute__, we provide a GASNETT_USE_* mechanism to force use
of a specific restrict keyword regardless of compiler version matching.