"Port" to Sun's "GCCFSS" (GCC for SPARC Systems)
This is actually work begun over 1 year ago when we first started trying
to test the UPC 2.8.0 release on the Niagra system(s) in the ParLab.
The "GCCFSS" compilers are the GNU front-ends grafted on to Sun's own
code generator for SPARC.

The main (so far to only) difference from a FSF GCC that I've observed is
that the "U" and "h" constraints are not recognized in inline asm.  We use
both of these in our 64-bit atomics for ILP32 on 64-bit SPARC CPUs.  So,
this commit adds a configure probe for support of "U" and "h" constraints.
At the moment the lack of these contraints triggers use of the mutex-based
64-bit atomics.  However, I am working on inline asm for this case that does
not require these contraints (at the expense of less flexibility for the
compiler to schedule regs, since we'll need to allocate fixed regs for some
purposes).