Report to user when $CC and $MPI_CC appear not to be the same compiler.
Since we are now better at supporting this case, we are have become LESS
likely to fail at compile time in the event of such a mismatch.  So, as a
service to the user we now warn them about the mismatch at configure time.
(of course we can't do anything about users who don't read the warning).