Implement --disable-debug-malloc (for bug 2731)
This commit allows one to disable GASNet's automatic use of MALLOC_CHECK_
at configure time.  This is a more robust work-around for bug 2731 (but
not a final fix) because it does not require any env var settings to
be made at run time.

NOTE: this only affects the automatic setting of MALLOC_CEHCK_ in the
GASNet tests Makefile.  Automatic setting of MALLOC_CHECK_ by upcrun
is STILL in place and will be addressed by another commit.