First attempt to fix bug 2786.
The reason of the try_all error is because DCMF_Alltoallv() doesn't copy the
data from the send buffer to the receive buffer when the receiver is the same
as the sender.  After adding the local data copy in gasnet dcmf conduit
exchange, try_all runs well without errors.

CVSMAIL: CC: