Add detection of the case where one has too many sockets for select().
As reported by Brad Chamberlain of Cray and Phil Nelson of Western
Washington University, a udp-conduit run w/ 63 nodes worked fine on
NetBSD5, but hung at exit time on 64 or more.  This has been reproduced
on FeeeBSD8 with an artificially reduced FD_SETSIZE value.  This commit
adds checking of the set size before the generation of the fd_set, which
will avoid the random memory write that would otherwise result.  If the
set is found to be too large, we abort since there is nothing else to do.
The set normally grows at startup as the slaves connect to the master
and thus the detection will take place at startup, rather than too late
at shutdown.

An assertion has also been added to the myselect() wrapper in case one
ever manages to find another path to select with an over-sized fd set.