Eliminate mutexes from PSHMNET queue pointer management.
The atomic CAS on the state field was already sufficient exclusion on the
recv path, and a change from read+write to CAS on the send path now provides
the exclusion.  This means that the q->next pointers can be advanced almost
immediately and multiple threads can perform meaningful overlap in both the
send and recv paths.

smp-conduit testam, testqueue and testcontend show measurable improvements
on both par and seq tests (though I am less than clear on why the seq case
improved).

There is still one mutex each on the send path (for buffer allocation) and
poll path (for round-robin over queues), but that is still better than the
two mutexes on each path prior to this commit.