Add additional types to the gasnete_anytype{8,16,32,64}_t unions.
Foremost this commit ensures that configure sets SIZEOF_{FLOAT,DOUBLE} in
gasnet_config.h to get these two standard types included in the unions.
Their accidental lack was noticed while chasing bug 2946.  It is possible
that their lack could account for some previously inexplicable bugs and
in particular ones that occur w/ non-gnu compilers.

Less importantly this commit adds the following types IF THEY EXIST:
 + "long double" which is an 8 byte type on at least Alpha and most PPC
 + "_Bool" which could be 1, 2, 4 or 8 bytes
 + "float _Complex" which is probably always 8 bytes
The existence check does not attempt to include complex.h or stdbool.h.

I've written, but commented out, configure logic to probe the SIZEOF_
 + "double _Complex"
 + "long double _Complex"
But since these are probably always 16 bytes or larger (if they are present
at all) they are not needed in the "anytype" unions.

I see we've not accounted for the possibility of a platform where
   sizeof(double) == sizeof(float)
I doubt that is an issue, but I wonder if some "insurance" might be in order.