Split vapi/ibv connection code into distinct functions in their own file. This is a MAJOR step toward partially connected operation (either statically connecting to a subset, or connectng dynamically).  However, a noted in the new file there are still some "global" assumptions that need fixing.
I tried the "trick" of cloning the RCS file to preserve the history of the
code as it had developed in gasnet_core.c.  I had hope that would work well
because the code stayed in the same order.  However, when I did try this I
found that 'cvs diff' was unable to identify the unchanged sections of code.
So, given that the trick could not do the one thing it was intended to, I
choose not to use it.