Fix ibv/vapi bug that broke any non-SRQ runs of either conduit last night. I introduced this problem in my connection work yesterday.
The gasnetc_sndrcv_init_peer() call must be made AFTER the QPs have reached
the INIT state or else the attempt to post RCV buffers is either a fatal
error (for VAPI) or is silently ignored (for IBV).  This was not noticed
in my SRQ-only testing because with SRQ is in use the RCV buffers are
posted to the SRQ instead of the QP.
I need to look more into the "silently ignored" to see if we are perhaps
just missing the error somehow.