Confine all direct knowledge of the dense cep table to gasnetc_init().
The storage is still dense, but outside of gasnetc_init() it is accessed
exclusively via the gasnetc_node2cep[] table (which has long been true in
the critical paths).  We also still depend on several things that are
initialized in that table at startup, which will need to be set "lazily"
when connections are dynamic.