Improved exit handling code in vapi/ibv conduit.
This code speeds the normal collective exit case at the expense of
slower exits in the non-collective cases.  This is acheived by
begining with a reduce-to-all (MAX) over exit codes, performed via AMs
and with a timeout.  If the reduction completes within the timeout
then we know not only the proper exit code for the job, but also know
that the exit is collective.  The reduction is done with
intra-supernode reduction followed by a reduction among
representatives from each supernode.  If the reduction timesout, then
we revert to the old code in which all nodes send an AM to 0 in order
to elect a "master" which then tries to initiate SIGQUIT exits on the
nodes that were not already in the exit path.

While this new code improves the performance of the exit, the
motivation was actaully to resolve an "issue" with the upcomming
support for partially-connected operation.  We will support a mode in
which you run (either fully-connected or dynamically
partialy-connected) and then dump the connection table for use in the
next run, in which the table is used at startup to establish only the
required connections.  The "issue" is that the old exit code included
a "race" (the election), which while doing no harm to the exit logic,
made the set of required connections unpredictable.  With the new code
any application with normal collective exit behavior has a
reproducable set of connections (assuming, of course, that the
application's own communication peerings are reproducable).

This commit also disables the cleanup of transport resources (the
various close/release/dealloc calls), which was not kept up-to-date
as support for SRQ and XRC were added, and which is proabably already
broken (we ignore error codes).  So, I may later remove this code.

I have also updates the README section which decribes the exit code.