Implement statically partially connected mode.
If env var GASNET_CONNECTFILE_IN is set then it is read to determine which
nodes we connect to, rather than all of them.

At the moment this only works if one disables the PSHM hybrid barrier which
has a randomizing property on the connection activity because the first notify
AM is sent by the last arrival (from among nodes sharing a supernode).
So, use of GASNET_CONNECTFILE_IN currently requires either a non-PSHM build,
or defining GASNETI_PSHM_BARRIER_HIER=0 (possibly in gasnet_extended_fwd.h).
I hope to modify the PSHM hybrid barrier to allow a compile-time change
of the last-arrival behavior to use of a specific represeantative.