Extend hybrid-PSHM AMDISSEM barrier to allow forcing fixed comms pattern. Also setup vapi/ibv conduit to request this modified behavior.
This disables sending of the initial Notify AM from the last arrival at the PSHM
barrier and instead forces it to always be sent from the single representative
of the supernode.  This is needed for the saving/restoring of connections,
or the restored conenctions could be "insufficient" to implement the barrier.
Next step is to do the same to the AMCENTRAL barrier, and then I hope to try
making this behavior runtime selectable so that vapi/ibv conduit can choose
this ONLY when saving or restoring the connection table.