Fix compilation problem w/ GASNet --enable-debug --disable-stats --disable-trace
The declaration and initialization of variable gasneti_mallocreport_filename
were all controlled by "#if GANETI_STATS_OR_TRACE", but the actual use was
instead controlled by "#if GASNET_DEBUG".  Since --enable-debug will also
enable stats and tracing by default, this went unnoticed in our testing.
Now the declaration, initialization and use of gasneti_mallocreport_filename
are all contrlled by "#if GASNET_DEBUG", since the malloc report is a
feature of the debugging malloc implementation and is NOT dependednt on the
presence of absence of stats and tracing support.
This compilation failure was a regression in GASNet 1.16.0, since in prior
releases the declaration was unconditional.  However, in all prior releases
the initialization was conditional, meaning that in a build with debugging
enabled and stats/tracing disabled the malloc report file would have been
unavailable.
Thanks to Nikita Andreev for reporting this issue.