Swap order of output flush calls. This quite reliably resolves the truncated/corrupted stdout and stderr that I've observed on NetBSD.  It seems almost certain that this represents some bug in NetBSD's libc, but I don't know anything specific.
Note that I did try adding a mutex to serialize the body of ...flush_streams(),
but that did not resolve the problems.  So, whatever the root cause might be
it is not a thread safetly problem.

This commit is made witout ertesting on every system with the belief that only
a buggy system (such as NetBSD apparently) should care about the order of these
flushes.  I will reformulate w/ appropriate #if's if we find that any
platforms get worse with this version of the flush code.