fix bug 3019 - cross-compile testtools fake conduit
The conduit is still faked as SMP in most cases, since that's usually the right choice when not cross-compiling.

On cross-compiled platforms, it fakes as MPI if and only if mpi-conduit is enabled (implying gasnet_mpirun is functional), and SMP otherwise (leaving the user to deal with test launch manually when the host machine cannot run target executables).
CVSMAIL: CC: