Defend against out-of-bounds return values from the portals library.
This is in response to the uts-upc failures seen on Franklin since
a s/w upgrade took place on that system in mid-September.
We are getting a portals-level error return, for a still unknown reason.
However, the problem was compounded by an out-of-bounds return value which,
when executing "ptl_err_str[retcode]", was yielding a random address which
in turn lead to a SIGSEGV when trying to print the error message.
Note that this does NOT fix the uts-upc runs, but at least now they are
dying with a (semi-)informative error message.