Oops, the "new" --enable-debug-malloc collides with an existing configure option.
This commit renames the seldom-used older option to --enable-system-debug-malloc.
This was the cause of "NEW" test failures on Carver today, because Carver has
one of the buggy glibc versions for which we automatically disabled use of
the system debug-malloc facility but an explict --enable-debug-malloc added
to the configure commandline defeated detection of the bad glibc.