Eliminate "thread tax" incurred by adding the connection thread. From the point-of-view of GASNet's weak atomics and semapores, the addition of the connection thread requires use of true atomic instructions.  While we DO need true atomics for the one semaphore and one LIFO that are accessed by both the connection thread AND the client's thread(s), we do NOT need them for the great majority of the internal data structures in a SEQ build w/ the connection thread.
This commit adds wrappers, with gasnetc_ prefix, for semaphores and
atomics which switch between sequential and concurrent
implementations based on the appropriate criteria, rather than
GASNETI_THREADS.

There is no longer ANY use of the weak atomics in vapi/ibv conduit.
We still use the "strong" atomics to get signal-safety in the exit
code, and in the shared memory file used for xrc setup.